+++ CHANGELOG.md
@@ -5,6 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [2.4.0] - 2023-09-29
+
+# Added
+
+Asyncio support. New Async Service classes are in /asyncio folder.
+
+
+## [2.3.0] - 2023-09-26
+
+# Added
+
+- FileScan Reversinglabs provider example
+- Domain WhoIs endpoint support
+
+# Changed
+
+- Deprecated config_id in PangeaConfig. Now is set in service initialization.
+
+# Fixed
+
+- HashType supported in File Intel
+
+
 ## [2.2.1] - 2023-09-06
 
 # Fixed
@@ -226,7 +250,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Domain Intel client
 - Redact client
 
-[unreleased]: https://github.com/pangeacyber/pangea-python/compare/v2.2.1...main
+[unreleased]: https://github.com/pangeacyber/pangea-python/compare/v2.4.0...main
+[2.4.0]: https://github.com/pangeacyber/pangea-python/compare/v2.4.0...v2.3.0
+[2.3.0]: https://github.com/pangeacyber/pangea-python/compare/v2.3.0...v2.2.1
 [2.2.1]: https://github.com/pangeacyber/pangea-python/compare/v2.2.1...v2.2.0
 [2.2.0]: https://github.com/pangeacyber/pangea-python/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/pangeacyber/pangea-python/compare/v2.0.0...v2.1.0